proper type

+++ alsa.c
@@ -79,7 +79,7 @@ max_samples_sio(void)
 }
 
 int16_t *
-read_sio(unsigned int n)
+read_sio(size_t n)
 {
 	snd_pcm_sframes_t rc;
